101. Symmetric Tree
